+++ src/controls/window.rs
@@ -152,6 +152,11 @@ impl Control for Window {
         ControlType::Window 
     }
 
+    fn children(&self) -> Vec<AnyHandle> {
+        use low::window_helper::list_window_children;
+        unsafe{ list_window_children(self.handle) }
+    }
+
     fn free(&mut self) {
         use user32::DestroyWindow;
         unsafe{ DestroyWindow(self.handle) };